+++ utilities/dristi-pdf/src/config.js
@@ -196,11 +196,6 @@ module.exports = {
       process.env.ORDER_ISSUE_WARRANT_QR || "order-issue-warrant-qr",
     order_issue_warrant:
       process.env.ORDER_ISSUE_WARRANT || "order-issue-warrant",
-    order_for_referral_case_Adr:
-      process.env.ORDER_FOR_REFERRAL_CASE_ADR || "order-for-referral-case-adr",
-    order_for_referral_case_Adr_qr:
-      process.env.ORDER_FOR_REFERRAL_CASE_ADR_QR ||
-      "order-for-referral-case-adr-qr",
   },
 
   app: {
