--- utilities/dristi-pdf/src/orderHandlers/orderForReferralCaseADR.js
@@ -1,208 +1,0 @@
-const cheerio = require("cheerio");
-const config = require("../config");
-const {
-  search_case,
-  search_order,
-  search_mdms,
-  search_hrms,
-  search_sunbirdrc_credential_service,
-  create_pdf,
-} = require("../api");
-const { renderError } = require("../utils/renderError");
-const { formatDate } = require("./formatDate");
-
-async function orderForReferralCaseADR(req, res, qrCode) {
-  const cnrNumber = req.query.cnrNumber;
-  const orderId = req.query.orderId;
-  const entityId = req.query.entityId;
-  const code = req.query.code;
-  const tenantId = req.query.tenantId;
-  const requestInfo = req.body.RequestInfo;
-
-  const missingFields = [];
-  if (!cnrNumber) missingFields.push("cnrNumber");
-  if (!orderId) missingFields.push("orderId");
-  if (!tenantId) missingFields.push("tenantId");
-  if (qrCode === "true" && (!entityId || !code))
-    missingFields.push("entityId and code");
-  if (requestInfo === undefined) missingFields.push("requestInfo");
-
-  if (missingFields.length > 0) {
-    return renderError(
-      res,
-      `${missingFields.join(", ")} are mandatory to generate the PDF`,
-      400
-    );
-  }
-
-  // Function to handle API calls
-  const handleApiCall = async (apiCall, errorMessage) => {
-    try {
-      return await apiCall();
-    } catch (ex) {
-      renderError(res, `${errorMessage}`, 500, ex);
-      throw ex; // Ensure the function stops on error
-    }
-  };
-
-  try {
-    // Search for case details
-    const resCase = await handleApiCall(
-      () => search_case(cnrNumber, tenantId, requestInfo),
-      "Failed to query case service"
-    );
-    const courtCase = resCase?.data?.criteria[0]?.responseList[0];
-    if (!courtCase) {
-      renderError(res, "Court case not found", 404);
-    }
-
-    // Search for HRMS details
-    // const resHrms = await handleApiCall(
-    //     () => search_hrms(tenantId, "JUDGE", courtCase.courtId, requestInfo),
-    //     "Failed to query HRMS service"
-    // );
-    // const employee = resHrms?.data?.Employees[0];
-    // if (!employee) {
-    //     renderError(res, "Employee not found", 404);
-    // }
-
-    // Search for MDMS court room details
-    const resMdms = await handleApiCall(
-      () =>
-        search_mdms(
-          courtCase.courtId,
-          "common-masters.Court_Rooms",
-          tenantId,
-          requestInfo
-        ),
-      "Failed to query MDMS service for court room"
-    );
-    const mdmsCourtRoom = resMdms?.data?.mdms[0]?.data;
-    if (!mdmsCourtRoom) {
-      renderError(res, "Court room MDMS master not found", 404);
-    }
-
-    // Search for MDMS court establishment details
-
-    // const resMdms1 = await handleApiCall(
-    //   () =>
-    //     search_mdms(
-    //       mdmsCourtRoom.courtEstablishmentId,
-    //       "case.CourtEstablishment",
-    //       tenantId,
-    //       requestInfo
-    //     ),
-    //   "Failed to query MDMS service for court establishment"
-    // );
-    // const mdmsCourtEstablishment = resMdms1?.data?.mdms[0]?.data;
-    // if (!mdmsCourtEstablishment) {
-    //   renderError(res, "Court establishment MDMS master not found", 404);
-    // }
-
-    // Search for order details
-    const resOrder = await handleApiCall(
-      () => search_order(tenantId, orderId, requestInfo),
-      "Failed to query order service"
-    );
-    const order = resOrder?.data?.list[0];
-    if (!order) {
-      renderError(res, "Order not found", 404);
-    }
-
-    // Handle QR code if enabled
-    let base64Url = "";
-    if (qrCode === "true") {
-      const resCredential = await handleApiCall(
-        () =>
-          search_sunbirdrc_credential_service(
-            tenantId,
-            code,
-            entityId,
-            requestInfo
-          ),
-        "Failed to query sunbirdrc credential service"
-      );
-      const $ = cheerio.load(resCredential.data);
-      const imgTag = $("img");
-      if (imgTag.length === 0) {
-        return renderError(
-          res,
-          "No img tag found in the sunbirdrc response",
-          500
-        );
-      }
-      base64Url = imgTag.attr("src");
-    }
-
-    let caseYear;
-    if (typeof courtCase.filingDate === "string") {
-      caseYear = courtCase.filingDate.slice(-4);
-    } else if (courtCase.filingDate instanceof Date) {
-      caseYear = courtCase.filingDate.getFullYear();
-    } else if (typeof courtCase.filingDate === "number") {
-      // Assuming the number is in milliseconds (epoch time)
-      caseYear = new Date(courtCase.filingDate).getFullYear();
-    } else {
-      return renderError(res, "Invalid filingDate format", 500);
-    }
-
-    const currentDate = new Date();
-    const formattedToday = formatDate(currentDate, "DD-MM-YYYY");
-    const additionalComments = order?.comments || "";
-    const modeOfAdr = order?.orderDetails?.adrMode || "";
-    // Prepare data for PDF generation
-    const data = {
-      Data: [
-        {
-          courtName: mdmsCourtRoom.name,
-          place: "Kollam",
-          state: "Kerala",
-          caseName: courtCase.caseTitle,
-          caseNumber: courtCase.caseNumber,
-          caseYear: caseYear,
-          modeOfAdr: modeOfAdr,
-          additionalComments: additionalComments,
-          date: formattedToday,
-          judgeSignature: "Judge Signature",
-          judgeName: "Suresh Soren",
-          courtSeal: "Court Seal",
-          qrCodeUrl: base64Url,
-        },
-      ],
-    };
-    console.log(data);
-
-    // Generate the PDF
-    const pdfKey =
-      qrCode === "true"
-        ? config.pdf.order_for_referral_case_Adr_qr
-        : config.pdf.order_for_referral_case_Adr;
-    const pdfResponse = await handleApiCall(
-      () => create_pdf(tenantId, pdfKey, data, req.body),
-      "Failed to generate PDF of referral case ADR"
-    );
-
-    const filename = `${pdfKey}_${new Date().getTime()}`;
-    res.writeHead(200, {
-      "Content-Type": "application/pdf",
-      "Content-Disposition": `attachment; filename=${filename}.pdf`,
-    });
-    pdfResponse.data
-      .pipe(res)
-      .on("finish", () => {
-        res.end();
-      })
-      .on("error", (err) => {
-        return renderError(res, "Failed to send PDF response", 500, err);
-      });
-  } catch (ex) {
-    return renderError(
-      res,
-      "Failed to query details of referral case ADR",
-      500,
-      ex
-    );
-  }
-}
-
-module.exports = orderForReferralCaseADR;
